Form submits to API and handles any errors
Moves subscription form code into a seperate file
Handles errors for unsuccessful submission
