+++ app/main/__init__.py
@@ -3,5 +3,5 @@ from flask import Blueprint
 main = Blueprint('main', __name__)  # noqa
 
 from app.main.views import (  # noqa
-    index, api, admin
+    index, api, admin, subscription
 )
