+++ app/templates/views/subscription.html
@@ -13,6 +13,9 @@
         {{ subscription_form.email(value=email) }}
         <button type="submit" class="btn btn-outline-success">Submit</button>
       </form>
+      {% if error %}
+        <li class="error_text">Failed to process email, please try again later</li>
+      {% endif %}
     </div>
     <div class="col-xs-12 col-sm-12 col-md-6">
       <iframe src="/static/pdfs/Website_Privacy_Notice.pdf" width="100%" height="480"></iframe>
